fix(pns): stop scalarProduct fusing the multiply-add on arm64
The PNS noise energy accumulator carried a comment saying the C reference
fuses its multiply-add, so the naive Go form was the matching translation
and a rounding guard must not be added. That was calibrated against the
darwin oracle, which is Apple clang and defaults to -ffp-contract=on.
Issue #15 established that the canonical reference is a contraction-off
build: gcc 14 at -std=c17 does not fuse, CI builds that way, and every
committed fixture was generated that way. So the comment had it backwards.

The consequence was worse than a rounding mismatch. gc fuses this loop
into FMADDS on arm64 and leaves MULSS+ADDSS on amd64, so the PNS band
energy, the NOISE_BT decision it feeds, and the emitted bitstream all
depended on GOARCH. The new fixture case below fails on arm64 and passes
on amd64 without this change, and passes on both with it.

Also extend tools/ctwoloop with fix_stereo_tools and a matching Go golden
test for the whole non-NMR stereo tool chain in aac_encode_frame order:
mark_pns, search_for_quantizers_twoloop, search_for_pns, search_for_is,
apply_intensity_stereo, search_for_ms, apply_mid_side_stereo.

The existing stereo fixture runs only long frames with PNS off, so it
never saw a NOISE_BT band, never saw an EIGHT_SHORT window group, and
never saw a band the quantizer had zeroed. Those are the inputs that make
the chain decide differently rather than merely round differently, which
is why a bit-exact fixture and a diverging end to end result could both
be true at once (issue #53). The new fixture records every discrete
output of the chain (can_pns, zeroes, max_sfb, band_type, sf_idx,
is_mask, ms_mask) next to the floats, and two of its six cases are tuned
to arm twoloop's "trade overdistorted bands for PNS-able zeroes" branch,
which no other fixture reached; that branch swaps coded content for
noise, so an error there is audible rather than subtle.

twoloop_search.bin and twoloop_stereo.bin are unchanged, byte for byte,
regenerated from the pinned tree to prove it.

+++ internal/coder/pns.go
@@ -218,14 +218,21 @@ func (c *Coder) SearchForPNS(sampleRate, bandwidth int,
 	}
 }
 
-// scalarProduct mirrors the archive's ff_scalarproduct_float_c
-// @ d09d5afc3a as clang compiles it on this platform: the multiply-add is
-// FMA-contracted, so the naive Go form (which gc also fuses on arm64) is
-// the matching translation. Do NOT add a rounding guard here.
+// scalarProduct mirrors ff_scalarproduct_float_c (libavutil/float_dsp.c
+// @ d09d5afc3a), whose accumulator is a plain float32 multiply followed by
+// an add.
+//
+// The float32 conversion is a rounding barrier and is load-bearing. Without
+// it gc fuses the multiply-add into FMADDS on arm64 but leaves MULSS+ADDSS
+// on amd64, which makes the PNS noise energy, and through it the NOISE_BT
+// decision and the emitted bitstream, depend on GOARCH. The reference is a
+// -ffp-contract=off build (the semantics every committed fixture was
+// generated under, see issue #15), which does not fuse either, so rounding
+// here is both the portable and the faithful choice.
 func scalarProduct(v1, v2 []float32) float32 {
 	p := float32(0.0)
 	for i := range v1 {
-		p += v1[i] * v2[i]
+		p += float32(v1[i] * v2[i]) // no cross-statement FMA
 	}
 	return p
 }
